Rename gen module to non-keyword
