+++ savvy-bindgen/src/codegen/static_files.rs
