+++ savvy-bindgen/src/codegen/templates/Cargo_toml
@@ -1,7 +1,7 @@
 [package]
 name = "{}"
 version = "0.1.0"
-edition = "2021"
+edition = "2024"
 
 [lib]
 crate-type = ["staticlib", "lib"]
@@ -11,10 +11,10 @@ crate-type = ["staticlib", "lib"]
 [profile.release]
 # By default, on release build, savvy terminates the R session when a panic
 # occurs. This is the right behavior in that a panic means such a fatal event
-# where we can have no hope of recovery. 
-# 
+# where we can have no hope of recovery.
+#
 # cf. https://doc.rust-lang.org/book/ch09-03-to-panic-or-not-to-panic.html
-# 
+#
 # However, it's possible that the panic is thrown by some of the dependency
 # crate and there's little you can do. In such cases, you can change the
 # following line to `panic = "unwind"` to always catch a panic.
