+++ savvy-bindgen/src/codegen/templates/Makevars.win.in
