+++ savvy-bindgen/src/codegen/templates/configure.win
