+++ savvy-bindgen/src/codegen/templates/dllname-win.def
