+++ savvy-bindgen/src/codegen/templates/gitignore
